Add Loki exporter, manually queue db transactions
Problem here is SQLite is locking up and sequelize is not doing a good job of retrying; but this is mostly happening in "prod" (i.e., my NAS). Having pulled the "prod" db and run locally, everything works as it should here.

I don't feel good about this solution, but I haven't had much time to work on this and I actually do *use* this application to manage my audiobooks, so I need to get it working again. Currently, imports fail and most cover requests too bc of db lock.